requirements.txt
langchain
langchain-openai
langchain-prolog
faiss-cpu
openai
tiktoken
typing-extensions
python-dotenv
